BUG FIX: Fixed Delete on Kill, SquadMember, Squad. BUG FIX: Added FetchType.LAZY to SquadMember, Squad, Kill

+++ src/main/java/com/example/human_vs_zombies/services/kill/KillServiceImpl.java
@@ -47,8 +47,9 @@ public class KillServiceImpl implements KillService {
     @Override
     public void deleteById(Integer id) {
 
-        killRepository.deleteById(id);
-
+        if(killRepository.existsById(id)){
+            killRepository.deleteById(id);
+        }
 
     }
 }
